refactor(fcc): shared orion.fcc.mcp_config for hub and harness
