+++ orion/fcc/__init__.py
@@ -1,0 +1,1 @@
+"""Shared FCC (agent-claude / harness) helpers."""
