+++ orion/fcc/fcc_claude_mcp.template.json
